pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "google-genai>=1.25.0",
    "google-generativeai>=0.8.5",
    "oauthlib>=3.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pdf2image>=1.17.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sift-stack-py>=0.7.0",
    "xlrd>=2.0.2",
]
